+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [0.1.0]
+* Subscriptions Delete
 * Subscriptions Create
 * Users Examples 1 - 5 https://docs.microsoft.com/en-us/graph/api/user-list
 * Initial Version